Setup basic project layout
This commit sets up the basic layout for the program. It transitions
from NodeJS to Bun, a more efficient alternative that promises enhanced
performance which is something very critical for the tool's
requirements.

Additionally I have integrated the yargs library to introduce a
rudimentary command line interface for the tool. The CLI currently
supports a singular command named `generate <filename>`, designed to
recieve a filename parameter listing the websites we want to generate a
report for. Here is a example of the contents of the file:

```
https://www.amazon.com/
https://www.google.com/
```

Upon execution, the crawler component of the program reads the filename,
and creates a folder for each website inside the `reports` folder. The
generated reports are then saved within their respective folders.